+++ src/pages/_document.tsx
@@ -4,6 +4,8 @@ import Document, { Html, Head, Main, NextScript } from "next/document";
 class MyDocument extends Document {
   static async getInitialProps(ctx) {
     const initialProps = await Document.getInitialProps(ctx);
+    // Invoke the newrelic agent
+    require("newrelic");
     return { ...initialProps };
   }