+++ playwright/tests/keywordSearch.spec.ts
@@ -11,7 +11,7 @@ Given(
 
 When("I input a search term", async function (this: CustomWorld) {
   await this.page.locator('[aria-label="Item Search"]').fill("IBM 1401");
-  await this.page.locator('[data-testid="button"]').click();
+  await this.page.locator("#searchbar-button-search-bar").click();
 });
 
 Then(