+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Upgrade to NYPL Design System version 1.1.0
 - Upgrade to React 17.0.2
 - Upgrade to Next 12.2.5
+- Fix undefined-heading id on details list heading bug
 
 ## [0.14.1]