+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Upgrade to React 17.0.2
 - Upgrade to Next 12.2.5
 - Fix undefined-heading id on details list heading bug
+- Selected language filters are displayed until removed/cleared
 
 ## [0.14.1]