+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Upgrade to React 17.0.2
 - Upgrade to Next 12.2.5
 - Fix undefined-heading id on details list heading bug
+- Add instrumentation for New Relic
 
 ## [0.14.1]