+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [Pre-release]
 
 - add features and tests for item details pages
+- add features and tests for EDD requests
+- add features and tests for item details pages
 - add features and tests for link targets
 - add features and tests for home page elements
 - add features and tests for keyword search